Prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,7 +5,7 @@
 */
 int main(void)
 {
-	unsigned long long int n;
+	unsigned long int n;
 
 	n = 612852475143;
 	int div = 2, ans = 0, maxFact;
